feat: refactor state logic
